// src/internal/instance/types.ts
export interface Instance {
  /** The Unique ID of the extension generated by nanoID */
  instanceID: string;
  /** A name choosen by the user. This name will be shown in the interface and searchable */
  name: string;
  /** The keyboard shortcut to fire up any extension (e.g CTRL+ALT+A) */
  keyboard?: string;
  /** The id of the linked extension */
  extensionID: string;
}
import { envVariables } from "../../internal/extension/types";
export interface updateEnvVariables extends envVariables {
  /** The value to update */
  value: string | number | boolean;
}